Migrated some MiCasa stuff in as a base from which to build a new client that will work in an essentially different manner from MiCasa (indeed MiCasa may get the boot once we have ToyBox because it's better).
With ToyBox the client is launched with a particular game configuration
and connects to the server and enters the lobby of that game
specifically. It knows nothing about other game lobbies and need not
provide a navigation mechanism between them. That would be handled on a
web site where a more sensible game directory with information for each
game could be provided.

Additionally, this frees up the games themselves to be hosted in a variety
of server (machine) configurations: all on one server or grouped onto
different servers in whatever way is appropriate.