Update the bootstrap to point to the public stuff.
Of course all THAT is probably going to move over to github at some point
here, but I'll ignore that for the moment.

Also, this involves a version bump from 1.2 to 2.1 which I haven't actually
tested & might break things. I'll give it all a whirl later, but I should
really get back to other things for now. (I don't feel TOO heartbroken
about leaving it in a broken state for now since it was already pointing
to a server the world at large can't reach)